pyproject.toml
[project]
name = "file-ingester"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "qdrant-client>=1.9",
  "sentence-transformers>=3.0",
  "docling>=2.0",
  "watchdog>=4.0",
  "apscheduler>=3.10",
  "sqlalchemy>=2.0",
  "aiosqlite>=0.20",
  "openpyxl>=3.1",
  "pandas>=2.0",
  "pygments>=2.18",
  "loguru>=0.7",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.11",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ingester*", "server*", "scheduler*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
